Add AutoFit Move Cifar10 example code into ImageClassification
