+++ test/runtests.jl
@@ -50,7 +50,10 @@ using Test
     @info "Testing README examples"
     include("examples.jl")
 
-    @info "Testing CIFAR10 example"
-    include("examples/Cifar10.jl")
+    @info "Testing autofit"
+    include("app/autofit.jl")
+
+    #@info "Testing CIFAR10 example"
+    #include("examples/Cifar10.jl")
 
 end
